requirements.txt
torch>=2.3.0
peft>=0.11
trl>=0.10.1
pyyaml>=6.0.2
click>=8.1.7
transformers>=4.44.2